Changed UI theme system to allow for changing of accents (for different office apps). Changing the base accent and moving the green accent into the excel VSTO project.
Changed approach to threading of UI and Execution Engines. PythonNET now cooperates with new threading approach.

Added Executor which can be passed to engines to allow for synchronous execution of code on the host thread, removing need for passing dispatchers or threads to the engine or ui.

Add small WinForms test project to ensure that execution engines are performing.